refactor(pages): move `Pages` to its own file
